Handle qgroup IDs matching subvolume IDs
The quota limits we will be showing for now are the limits of the
subvolume's default quota group, that is "0/<subvol id>". One can
add more qgroups to the file system with a higher level than 0 and
an arbitrary second number. However, if that second number matches
the subvolume ID of a given subvolume, the quota info for that subvolume
was overwritten by the higher-level qgroup (independent of any
child/parent relationship between the two quota groups).

This commit skips all quota groups with a level other than 0.